Add MovementSpeedPowerUp to the Game.
Upon rendering a map, there are now five MovementSpeedPowerUps placed beneath a random DestructibleWall, given that there are no other PowerUps under this wall.
The maximum movement seed of the Player is set to 4.0 (starting from 2.0). Each power-up increases the movement speed by 0.5.
The current movement speed is displayed in the Hud.